+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-## Indev 1.0.1 - Nov 30, 2018 Tuesday
+## Indev 1.1.0 - Dec 1, 2018
+- [ADDITION]
+  - [PLAYER] Added player starting position
+
+## Indev 1.0.1 - Nov 30, 2018
 - [FIX] [GRAPHIC] The display problem (at ofs: 10, 10) crashes fixed :)
 
-## Indev 1.0a - Nov 27, 2018 Tuesday
+## Indev 1.0a - Nov 27, 2018
 - Update Text File
 
-## Indev 1.0 - Nov 27, 2018 Tuesday
+## Indev 1.0 - Nov 27, 2018
 - Initial Commit
 - NOTE: Project already started